pyproject.toml
[project]
name = "memesmithy-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "easyocr>=1.7.2",
    "fire>=0.7.0",
    "httpx>=0.28.1",
    "litellm>=1.65.1",
    "matplotlib>=3.10.1",
    "mplcursors>=0.6",
    "nextcord>=3.0.1",
    "openai>=1.70.0",
    "opencv-python>=4.11.0.86",
    "python-doctr>=0.11.0",
    "python-dotenv>=1.1.0",
    "smolagents>=1.13.0",
    "supabase>=2.15.0",
]

[dependency-groups]
dev = [
    "autopep8>=2.3.2",
]
